Move logic from utils.js to sketches/index.js
