+++ src/activities/creatureCreator/sketches/index.js
@@ -1,5 +1,14 @@
-import { CANVAS_WIDTH, CANVAS_HEIGHT, CANVAS_BG_COLOR } from "../utils";
+/**
+ * CONSTANTS
+ */
+const CANVAS_WIDTH = 200;
+const CENTER_X = CANVAS_WIDTH / 2;
+const CANVAS_HEIGHT = 200;
+const CANVAS_BG_COLOR = [243, 243, 243];
 
+/**
+ * P5
+ */
 const sketch = p5 => {
   let eyes;
 
@@ -24,11 +33,27 @@ const sketch = p5 => {
     });
   };
 
-  p5.setCreature = props => {
+  p5.setCreature = type => {
     p5.clear();
-    eyes = props.eyes;
+    eyes = generateEyes();
     draw();
   };
 };
 
 export default sketch;
+
+/**
+ * HELPER FUNCTIONS
+ */
+const generateEyes = () => {
+  const eyeSize = randomInt(20, 50);
+  const leftEyeX = randomInt(eyeSize, CENTER_X - eyeSize);
+  const rightEyeX = CANVAS_WIDTH - (CENTER_X - leftEyeX);
+  const yPos = randomInt(eyeSize, CANVAS_HEIGHT / 2 - eyeSize);
+
+  return [[leftEyeX, yPos, eyeSize], [rightEyeX, yPos, eyeSize]];
+};
+
+const randomInt = (min, max) => {
+  return Math.floor(Math.random() * Math.floor(max + 1));
+};
